Fix interfaces in aggregation filters
